+++ mobile/android/android-components/components/browser/session/src/main/java/mozilla/components/browser/session/engine/EngineObserver.kt
@@ -19,6 +19,7 @@ import mozilla.components.browser.state.action.ContentAction
 import mozilla.components.browser.state.action.CrashAction
 import mozilla.components.browser.state.action.EngineAction
 import mozilla.components.browser.state.action.MediaAction
+import mozilla.components.browser.state.action.MediaSessionAction
 import mozilla.components.browser.state.action.TrackingProtectionAction
 import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.state.state.content.DownloadState
@@ -32,6 +33,7 @@ import mozilla.components.concept.engine.history.HistoryItem
 import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.concept.engine.media.Media
 import mozilla.components.concept.engine.media.RecordingDevice
+import mozilla.components.concept.engine.mediasession.MediaSession
 import mozilla.components.concept.engine.permission.PermissionRequest
 import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.concept.engine.window.WindowRequest
@@ -306,6 +308,60 @@ internal class EngineObserver(
         }
 
         mediaMap.remove(media)
+    }
+
+    override fun onMediaActivated(mediaSessionController: MediaSession.Controller) {
+        store?.dispatch(MediaSessionAction.ActivatedMediaSessionAction(
+            session.id,
+            mediaSessionController
+        ))
+    }
+
+    override fun onMediaDeactivated() {
+        store?.dispatch(MediaSessionAction.DeactivatedMediaSessionAction(session.id))
+    }
+
+    override fun onMediaMetadataChanged(metadata: MediaSession.Metadata) {
+        store?.dispatch(MediaSessionAction.UpdateMediaMetadataAction(session.id, metadata))
+    }
+
+    override fun onMediaPlaybackStateChanged(playbackState: MediaSession.PlaybackState) {
+        store?.dispatch(MediaSessionAction.UpdateMediaPlaybackStateAction(
+            session.id,
+            playbackState
+        ))
+    }
+
+    override fun onMediaFeatureChanged(features: MediaSession.Feature) {
+        store?.dispatch(MediaSessionAction.UpdateMediaFeatureAction(
+            session.id,
+            features
+        ))
+    }
+
+    override fun onMediaPositionStateChanged(positionState: MediaSession.PositionState) {
+        store?.dispatch(MediaSessionAction.UpdateMediaPositionStateAction(
+            session.id,
+            positionState
+        ))
+    }
+
+    override fun onMediaMuteChanged(muted: Boolean) {
+        store?.dispatch(MediaSessionAction.UpdateMediaMutedAction(
+            session.id,
+            muted
+        ))
+    }
+
+    override fun onMediaFullscreenChanged(
+        fullscreen: Boolean,
+        elementMetadata: MediaSession.ElementMetadata?
+    ) {
+        store?.dispatch(MediaSessionAction.UpdateMediaFullscreenAction(
+            session.id,
+            fullscreen,
+            elementMetadata
+        ))
     }
 
     override fun onWebAppManifestLoaded(manifest: WebAppManifest) {