+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/action/BrowserAction.kt
@@ -14,6 +14,7 @@ import mozilla.components.browser.state.state.ContainerState
 import mozilla.components.browser.state.state.ContentState
 import mozilla.components.browser.state.state.CustomTabSessionState
 import mozilla.components.browser.state.state.EngineState
+import mozilla.components.browser.state.state.MediaSessionState
 import mozilla.components.browser.state.state.MediaState
 import mozilla.components.browser.state.state.ReaderState
 import mozilla.components.browser.state.state.SecurityInfoState
@@ -34,6 +35,7 @@ import mozilla.components.concept.engine.content.blocking.Tracker
 import mozilla.components.concept.engine.history.HistoryItem
 import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.concept.engine.media.Media
+import mozilla.components.concept.engine.mediasession.MediaSession
 import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.concept.engine.search.SearchRequest
 import mozilla.components.concept.engine.webextension.WebExtensionBrowserAction
@@ -819,6 +821,75 @@ sealed class MediaAction : BrowserAction() {
     data class UpdateMediaAggregateAction(
         val aggregate: MediaState.Aggregate
     ) : MediaAction()
+}
+
+/**
+ * [BrowserAction] implementations related to updating the [MediaSessionState].
+ */
+sealed class MediaSessionAction : BrowserAction() {
+    /**
+     * Activates [MediaSession] owned by the tab with id [tabId].
+     */
+    data class ActivatedMediaSessionAction(
+        val tabId: String,
+        val mediaSessionController: MediaSession.Controller
+    ) : MediaSessionAction()
+
+    /**
+     * Activates [MediaSession] owned by the tab with id [tabId].
+     */
+    data class DeactivatedMediaSessionAction(
+        val tabId: String
+    ) : MediaSessionAction()
+
+    /**
+     * Updates the [MediaSession.Metadata] owned by the tab with id [tabId].
+     */
+    data class UpdateMediaMetadataAction(
+        val tabId: String,
+        val metadata: MediaSession.Metadata
+    ) : MediaSessionAction()
+
+    /**
+     * Updates the [MediaSession.PlaybackState] owned by the tab with id [tabId].
+     */
+    data class UpdateMediaPlaybackStateAction(
+        val tabId: String,
+        val playbackState: MediaSession.PlaybackState
+    ) : MediaSessionAction()
+
+    /**
+     * Updates the [MediaSession.Feature] owned by the tab with id [tabId].
+     */
+    data class UpdateMediaFeatureAction(
+        val tabId: String,
+        val features: MediaSession.Feature
+    ) : MediaSessionAction()
+
+    /**
+     * Updates the [MediaSession.PositionState] owned by the tab with id [tabId].
+     */
+    data class UpdateMediaPositionStateAction(
+        val tabId: String,
+        val positionState: MediaSession.PositionState
+    ) : MediaSessionAction()
+
+    /**
+     * Updates the [muted] owned by the tab with id [tabId].
+     */
+    data class UpdateMediaMutedAction(
+        val tabId: String,
+        val muted: Boolean
+    ) : MediaSessionAction()
+
+    /**
+     * Updates the [fullScreen] and [MediaSession.ElementMetadata] owned by the tab with id [tabId].
+     */
+    data class UpdateMediaFullscreenAction(
+        val tabId: String,
+        val fullScreen: Boolean,
+        val elementMetadata: MediaSession.ElementMetadata?
+    ) : MediaSessionAction()
 }
 
 /**